fix: duolicate define
