+++ PCL.Core/Minecraft/Profile/Autnenrication/IAuthenticateProvider.cs
@@ -1,0 +1,8 @@
+﻿using System.Threading.Tasks;
+
+namespace PCL.Core.Minecraft.Profile.Autnenrication;
+
+public interface IAuthenticateProvider
+{
+    //public Task AuthenticateAsync();
+}
