+++ PCL.Core/Minecraft/Profile/Autnenrication/MicrosoftProvider.cs
@@ -1,0 +1,6 @@
+﻿namespace PCL.Core.Minecraft.Profile.Autnenrication;
+
+public class MicrosoftProvider
+{
+    
+}
