+++ PCL.Core/Minecraft/Profile/Models/ProfileType.cs
@@ -1,0 +1,8 @@
+﻿namespace PCL.Core.Minecraft.Profile.Models;
+
+public enum ProfileType
+{
+    Microsoft,
+    Offline,
+    Authlib
+}
