--- PCL.Core/Minecraft/Profile/Profile/Autnenrication/MicrosoftProvider.cs
@@ -1,6 +1,0 @@
-﻿namespace PCL.Core.Minecraft.Profile.Autnenrication;
-
-public class MicrosoftProvider
-{
-    
-}
