--- PCL.Core/Minecraft/Profile/Profile/Autnenrication/Utils/Models/XboxProperty.cs
@@ -1,8 +1,0 @@
-﻿namespace PCL.Core.Minecraft.Profile.Autnenrication.Utils.Models;
-
-public record XboxProperty
-{
-    public string AuthMethod { get; set; } = "RPS";
-    public string SiteName { get; set; } = "user.auth.xboxlive.com";
-    public required string RpsTicket { get; set; }
-}
