--- PCL.Core/Minecraft/Profile/Profile/Autnenrication/Utils/MojangUtils.cs
@@ -1,8 +1,0 @@
-﻿namespace PCL.Core.Minecraft.Profile.Autnenrication.Utils;
-
-public static class MojangUtils
-{
-    private const string MojangAuthServer = "https://api.minecraftservices.com/authentication/login_with_xbox";
-    //public static async Task<> AuthenticateAsync(){}
-    //public static async Task<bool>
-}
