--- PCL.Core/Minecraft/Profile/Profile/Models/ProfileJson.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections.Immutable;
-
-namespace PCL.Core.Minecraft.Profile.Models;
-
-public class ProfileJson<T> where T : class
-{
-    public int LastUsed { get; set; }
-    public ImmutableArray<T> Profiles { get; private set; } = [];
-}
