--- PCL.Core/Minecraft/Profile/Profile/Models/ProfileType.cs
@@ -1,8 +1,0 @@
-﻿namespace PCL.Core.Minecraft.Profile.Models;
-
-public enum ProfileType
-{
-    Microsoft,
-    Offline,
-    Authlib
-}
